recipes-bsp: imx-atf: Fix UART4 crash
UART4 is assigned to the Cortex-M4 in the RDC by default in
the imx-atf package. UART4 is needed by axon-imx8mm to
communicate with the bluetooth radio from the Cortex-A53
domain. This patch adds a patch to make this assignment
in the ATF.

The requirement for UART4 to be accessible by the A-cores
is specific to AXON-IMX8MM, where we use UART4 for Bluetooth
functionality. Set COMPATIBLE_MACHINES to axon-imx8mm.