recipes-devtools:python:python3-pybind11: add bbappend
When building imx-image-full (such as when attempting to build with meta-ml packages
the build will fail when building python3-pybind stating that it cannot find
_sysconfigdata:

ModuleNotFoundError: No module named '_sysconfigdata'

In reading the NXP community forum, it appears that the reason for the failure
is that the python executable is pointing to the Ubuntu 20.04 version (3.8.5)
and the needed version is the native tool (3.7). This bbappend adds to
EXTRA_OECMAKE to point it to the native tool path.

It should be noted that NXP in future revisions (gatesgarth) does not have a
separate recipe for python3-pybind11. It seems that the dependency on version 2.5.0
does not exist in gatesgarth.