recipes-devtools:python:python3-pybind11: Revert "recipes-devtools:python:python3-pybind11: add bbappend"
This reverts commit 5e992afffaf9bd075a5c8df72a2533ebd209579a.
This patch provides a workaround to fix the build failure in Yocto 3.0(Zeus).
But the dependency on version 2.5.0 does not exist in gatesgarth.
So we remove this patch in Yocto 3.2(gatesgarth).

Or it causes build failure as follows:
--------------------------------------------------------------------------------
-- Found PythonInterp: /yocto_3.2_gatesgarth_5.10.9-1.0.0/build-wayland-edm-g-imx8mp/tmp/work/cortexa53-crypto-poky-linux/python3-pybind11/2.5.0-r0/recipe-sysroot-native/usr/bin/python3-native/python3.7
CMake Error at tools/FindPythonLibsNew.cmake:96 (message):
  Python config failure:

Call Stack (most recent call first):
  tools/pybind11Tools.cmake:16 (find_package)
  CMakeLists.txt:33 (include)
--------------------------------------------------------------------------------