recipes-connectivity: openssh: don't create openssh-dev
populate_sdk installs all -dev packages, so having openssh-dev
pulls in openssh-sshd in as a dependency. That then conflicts with
Dropbear, which is only using openssh-sftp.

The openssh-dev package is empty and unused, created by bitbake.conf

It avoids the issue when populating SDK:

Error message:
------------------------------------------------------------------
The following packages have unmet dependencies:
dropbear : Conflicts: openssh but 8.7p1-r0 is to be installed
openssh : Depends: openssh-sshd but it is not installable
           Conflicts: dropbear but 2020.81-r0 is to be installed
E: Unable to correct problems, you have held broken packages.
------------------------------------------------------------------

Refer to:
https://github.com/openbmc/openbmc/commit/1467cd8872d1977b6a9a4bfb0c2dff9572a1d2df