tn-base: expand BSPDIR in python
`BSPDIR` is not a common Yocto construct and restricts how projects can
be set up. This patch expands the BSPDIR variable in Python which avoids
Poky delaying expansion if BSPDIR is not defined. The full solution is
likely to introduce a toolchain recommendation based on the presence of
some machine, distro, or image features value.